Expose core pipeline as library
